chore(travis): enable musl builds
Added x86_64-unknown-linux-musl builds for all the tool chains.
Enable `sudo: true` for musl builds on Travis CI so that `musl-tools` can be
installed with `apt-get`.
Set `dist: trusty` for musl Travis CI builds to build on Ubuntu Trusty.
Modified the `ci/install.sh` script to install the `musl-tool` package for the
`x86_64-unknown-linux-musl` target so that `musl-gcc` is available to certain
crates (e.g. `miniz-sys`).