feat(backend): move to glutin backend
Cut version 0.1.0.
Update clippy to 0.0.68.
Add cgmath, glium, glium_text, glutin, and image as dependencies.
Remove piston2d-graphics and piston2d-opengl_graphics as dependencies.

Remove `i686-pc-windows-msvc` and `x86_64-pc-windows-msvc` from supported
platforms list.
Add `appveyor.yml` configuration for building on `i686-pc-windows-gnu` and
`x86_64-pc-windows-gnu`.

Add localization strings for `internal_failed_to_load_image` and
`internal_failed_to_load_texture` messages.

Add the `block.ai` file for the new isometric block textures as well as a
`block.png` with the rasterized equivalent in PNG format.
Remove individual texture files `grass.png`, `sand.png`, `soil.png`, `wall.png`,
and `water.png`.

Add a `sprite` module to provide sprite rendering functionality to
`rgframework`.
Replace backend abstraction code in the `graphics` and `backend` modules of
`rgframework` with a module providing a `RenderContext` struct to store all
items which are relevant to rendering images to the screen (currently just the
glium `Context`).
Add a `Color` struct to `rgframework` for storing an RGBA color value.
Remove the `Draw` trait from `rgframework`.
Enable the `slice_patterns` feature in `rgframework`.
Add a `Manager` struct to `rgframework` for storing references to resources
using the `Weak` reference type, so that resources are stored only exactly as
long as they are needed, and only lazily loaded when they're necessary.
Add a `Rectangle` struct for representing a rectangle based on an origin point,
a width value, and a height value.
Add a `Mesh` struct, which is used to generate a mesh from a list of vertices
and faces, which can then be rendered.
Add a module to store a `Renderer` struct, which is used to render a `Mesh` on a
supplied surface, with the appropriate uniforms.
Add a `Renderable` implementation for the `Sprite` struct, allowing it to be
rendered by supplying a `Sprite` with a `Renderer`, a `Surface` and a parent
`Matrix4`.
Add a `TextRenderer` struct, which is similar to the `Renderer` struct, but is
exclusively used for rendering text to the screen.
Add a `Vertex` struct which stores a set of position and tex_coord values,
specifying a vertex which can then be used to generate a `Mesh`.
Modify the `Scene` and `SceneManager` structs to accept generic parameters for
`GenericEvent` and `Surface`.
Add a `Texture` struct which is used for loading textures, which can then be
rendered by providing them as uniform values.
Add a `Transform` struct, which provides the ability to compute a transformation
matrix by supplying values for position, scale, and anchor.
Add `opengl.png` for use in testing texture rendering.

Add a `matrix` method to the `Camera` struct to provide a `Matrix4` for use in
rendering.
Modify the `Game` struct to store a `Rc<SceneContext<F>>` instead of an
`Rc<Config>` and an `Rc<Localization>`.
Modify the `Game` struct to render using a glium `Context` instead of
`GlGraphics`.
Add a counter for the frame end time to be rendered as part of the window
information.
Add a debug information output to print out the `hidpi_factor` value for the
created window when loading the game.
Remove `glyph_cache` creation and initialization.
Add `text_renderer` creation and initialization.
Add a `texture_manager` `Manager` instance for loading and managing textures.
Add a vertex and fragment shader to the program as static strings used to
initialize the renderer.
Add a `SceneContext` struct to store values necessary for rendering scenes (e.g.
config, localization, etc.).
Add `BACKGROUND_COLOR`, `TEXT_COLOR`, `DISPLAY_SIZE`, `TILE_SIZE`, `TILE_COLOR`,
and `MARGIN` constants to the `game_scene` module.
Remove the `CAMERA_INITIAL_POSITION` and `CAMERA_MOVEMENT_SPEED` constants from
the `game_scene` module.
Re-write text rendering code to use the new `TextRenderer` in the `GameScene`
render method.
Re-write all non-text rendering code to use the new `Renderable` trait and
render using the rgframework `Renderer`.
Add `BACKGROUND_COLOR` and `TEXT_COLOR` to the `menu_scene` module.
Remove the `textures` module for loading OpenGL textures.
Add a `tile` module for storing information regarding the texture offsets for
the terrain `TileType` enum values.

Closes #88.
Breaking change: heavily modifies rendering code in rgframework.