feat: Add a state to the dwarf depending on if they're in freefall
Attach `user_data` to colliders so that we can look them up without
getting an `CannotReadArchetype` error.

Look up dwarves and planes when a contact is detected. If the dwarf
has entered into contact with the ground, they're either hit the
ground and become stationary or they've hit the ground but will bounce
again. We can detect when they're in contact with the ground for now.