fix: Render only the top transparent mesh per chunk
Render only the top transparent mesh of a chunk. This ensures that water
beneath the top layer is invisible to the player. Unfortunately this has
the side effect that the water is now one block thick and has no water
on the sides of the map, but there doesn't seem to be a way around that
for now.