feat: Use shaders to slice terrain
Add a custom vertex and fragment shader for the terrain meshes. This
allows us to slice the terrain based on a constant y-level. This
unfortunately makes the top of the mesh see-through, and removes all
lighting, etc. from the mesh. It also fails to run on WASM, but it's a
first step.