impr: Render terrain slicing with PBR
Copy over the bevy_pbr crate and add modifications to use an extra
YLevel uniform. This uniform is used to discard fragments above a
certain y-level. There's some glitches when slicing beneath the sea
level (z-fighting?) but aside from that it works.

I still need to try rendering the meshes that act as "hats" for the
chunks on-the-fly instead of caching them. That might not be fast enough
in the browser.