fix: Get sliced terrain shaders working on WASM
Since we use a custom fragment shader to slice the terrain, we need to
add a uniform to that shader. This means that we need to include and
override the bevy_webgl2 WASM fragment and vertex shaders.

When including just a float in the uniform, it doesn't seem to work due
to a bug in WebGL2:
https://web.archive.org/web/20210103080711/https://bugzilla.mozilla.org/show_bug.cgi?id=1569926
In the logs, I see:

    Error: WebGL warning: drawElements: Buffer for uniform block is smaller than UNIFORM_BLOCK_DATA_SIZE.

To get around this bug, I went with the simple route of declaring a vec3
and passing the y-level in through there. For whatever reason the code I
based this on allocates an `[f32; 4]` when providing a vec3 even though
an `[f32; 3]` should suffice. I'm too tired to diagnose this so I just
went with the oversized buffer.

Add copyright headers to files from other projects.
Add notice regarding modified works.

Remove unused files.