fix: Force GLSL to use set 4, binding 4 for YLevel uniform
Force set 4, binding 4 for the YLevel uniform. With the previous value,
it'd work in Firefox but in Chrome I'd see a GL_INVALID_OPERATION:

    It is undefined behaviour to have a used but unbound uniform buffer.

Removing the set, binding on the field entirely didn't work
because then Chrome would default to 4, 4 (which would work) and FF
would default to 0, 0 (IIRC; which wouldn't work).

Setting it to 4, 4 makes both browsers work.